Created a component for SkillBadges
