+++ client-web/src/components/SkillBadgeList.tsx
@@ -1,0 +1,65 @@
+import { useState } from "react";
+
+interface IProps {
+  skills: string[];
+  setSkills: (skills: string[]) => void;
+}
+
+export const SkillBadgeList = ({ skills, setSkills }: IProps) => {
+  const [newSkill, setNewSkill] = useState("");
+
+  const handleAddSkill = () => {
+    const skill = newSkill.trim();
+    if (skill && !skills?.includes(skill)) {
+      setSkills([...(skills ?? []), skill]);
+      setNewSkill("");
+    }
+  };
+
+  const updateSkills = (name: string) => {
+    const newList = skills?.filter((eachSkill) => eachSkill !== name);
+    setSkills(newList ?? []);
+  };
+  return (
+    <div className="flex items-center">
+      <div className="mr-2">Skills:</div>
+      <div>
+        {skills?.map((name) => (
+          <button
+            onClick={() => updateSkills(name)}
+            className="btn btn-outline btn-accent btn-sm mr-2 mb-2"
+            key={name}
+          >
+            {name.replace(/\b\w/g, (char) => char.toUpperCase())}
+            <span className="ml-1">&times;</span>
+          </button>
+        ))}
+        {skills?.length < 10 && (
+          <div className="flex mt-2">
+            <input
+              type="text"
+              className="input input-bordered mr-2"
+              placeholder="Add skill"
+              value={newSkill}
+              onChange={(e) => setNewSkill(e.target.value)}
+              onKeyDown={(e) => {
+                if (e.key === "Enter") {
+                  e.preventDefault();
+                  handleAddSkill();
+                }
+              }}
+            />
+            <button
+              type="button"
+              className="btn btn-success"
+              onClick={handleAddSkill}
+              disabled={!newSkill.trim() || skills?.includes(newSkill.trim())}
+            >
+              Add
+            </button>
+          </div>
+        )}
+      </div>
+    </div>
+  );
+};
